Theme customization moved from shop to theme
