+++ packages/Webkul/FPC/src/Listeners/ThemeCustomization.php
@@ -3,7 +3,7 @@
 namespace Webkul\FPC\Listeners;
 
 use Spatie\ResponseCache\Facades\ResponseCache;
-use Webkul\Shop\Repositories\ThemeCustomizationRepository;
+use Webkul\Theme\Repositories\ThemeCustomizationRepository;
 
 class ThemeCustomization
 {
